+++ libft/srcs/lft/ft_is_one_of_n.c
@@ -6,13 +6,13 @@
 /*   By: tmaluh <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/02/20 08:33:45 by tmaluh            #+#    #+#             */
-/*   Updated: 2019/04/09 23:13:21 by tmaluh           ###   ########.fr       */
+/*   Updated: 2019/05/12 14:33:45 by tmaluh           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-bool	ft_is_one_of_n(int64_t cmp, int64_t n, ...)
+bool	ft_is_one_of_n(int64_t cmp, size_t n, ...)
 {
 	va_list	ap;
 	bool	is_one_of;
